Move CameraPoseEstimator to AprilTagVisionIO so it can be used in sim
