+++ src/main/java/frc/robot/subsystems/apriltagvision/AprilTagVisionIOReal.java
@@ -1,22 +1,16 @@
 package frc.robot.subsystems.apriltagvision;
 
-import static frc.robot.Constants.AprilTagConstants.aprilTagFieldLayout;
 import static java.lang.System.arraycopy;
 
 import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.geometry.Pose3d;
-import edu.wpi.first.math.geometry.Transform3d;
 import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.wpilibj.Timer;
 import frc.robot.Constants.AprilTagConstants.CameraResolution;
 import java.util.AbstractMap.SimpleEntry;
-import java.util.List;
 import java.util.Optional;
 import org.photonvision.EstimatedRobotPose;
-import org.photonvision.PhotonCamera;
-import org.photonvision.PhotonPoseEstimator;
-import org.photonvision.PhotonPoseEstimator.PoseStrategy;
 import org.photonvision.targeting.PhotonPipelineResult;
 
 public class AprilTagVisionIOReal implements AprilTagVisionIO {
@@ -74,63 +68,6 @@ public class AprilTagVisionIOReal implements AprilTagVisionIO {
             timestampArray[index] = 0.0;
             latencyArray[index] = 0.0;
           });
-    }
-  }
-
-  /**
-   * A class containing a {@link PhotonCamera} and a {@link PhotonPoseEstimator} to keep the two
-   * objects together.
-   */
-  public static final class CameraPoseEstimator {
-    public final PhotonCamera camera;
-    public final Transform3d robotToCamera;
-
-    public final PhotonPoseEstimator poseEstimator;
-
-    /**
-     * A class containing a {@link PhotonCamera} and a {@link PhotonPoseEstimator} to keep the two
-     * objects together.
-     *
-     * @param name The name of the camera.
-     * @param robotToCamera The offset from the center of the robot to the camera.
-     * @param poseStrategy The {@link PoseStrategy} to use for the {@link PhotonPoseEstimator}.
-     */
-    public CameraPoseEstimator(
-        PhotonCamera camera,
-        Transform3d robotToCamera,
-        PhotonPoseEstimator.PoseStrategy poseStrategy) {
-      this.camera = camera;
-      this.robotToCamera = robotToCamera;
-
-      this.poseEstimator =
-          new PhotonPoseEstimator(aprilTagFieldLayout, poseStrategy, robotToCamera);
-    }
-
-    /**
-     * Updates the {@link PhotonPoseEstimator} by retrieving the unread results from the camera and
-     * estimating the robot's pose based on the latest result.
-     *
-     * @return A {@link SimpleEntry} containing: <br>
-     *     <br>
-     *     - An {@link Optional} of {@link EstimatedRobotPose} if the pose estimation is successful,
-     *     otherwise an empty {@link Optional}. (Accessible via {@link SimpleEntry}.getKey()) <br>
-     *     </br> - An {@link Optional} of {@link PhotonPipelineResult} containing the latest result
-     *     from the camera, or an empty {@link Optional} if there are no unread results. (Accessible
-     *     via {@link SimpleEntry}.getValue())
-     */
-    public SimpleEntry<Optional<EstimatedRobotPose>, Optional<PhotonPipelineResult>> update() {
-      List<PhotonPipelineResult> results = camera.getAllUnreadResults();
-
-      Optional<PhotonPipelineResult> result =
-          results.isEmpty()
-              ? Optional.empty()
-              : Optional.of(
-                  results.get(
-                      results.size()
-                          - 1)); // TODO Make sure this is the latest result, rather than oldest
-
-      return new SimpleEntry<Optional<EstimatedRobotPose>, Optional<PhotonPipelineResult>>(
-          result.flatMap(poseEstimator::update), result);
     }
   }
 }
